refactor: convert literal to constan
