--- src/main/java/com/yourssu/roomescape/auth/Role.java
@@ -1,5 +1,0 @@
-package com.yourssu.roomescape.auth;
-
-public class Role {
-    public static final String ADMIN = "ADMIN";
-}
