+++ src/main/java/com/yourssu/roomescape/constants/CookieConstants.java
@@ -1,0 +1,5 @@
+package com.yourssu.roomescape.constants;
+
+public class CookieConstants {
+    public static final String TOKEN_NAME = "token";
+}
