+++ src/main/java/com/yourssu/roomescape/constants/RoleConstants.java
@@ -1,0 +1,6 @@
+package com.yourssu.roomescape.constants;
+
+public class RoleConstants {
+    public static final String ADMIN = "ADMIN";
+    public static final String USER = "USER";
+}
