+++ src/main/java/com/yourssu/roomescape/security/LoginMember.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.auth;
+package com.yourssu.roomescape.auth.security;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
